backend/src/spatx_core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spatx_core"
version = "0.1.0"
description = "A modular spatial transcriptomics analysis framework"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Purushottam", email = "purushottam.tiwari.eee21@iitbhu.ac.in"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics"
]
dependencies = [
    "einops>=0.8.1",
    "pandas>=2.3.1",
    "scipy>=1.16.1",
    "timm>=1.0.19",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",  # for image processing
    "tqdm>=4.65.0",    # for progress bars
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",  # for coverage reporting
]
cu128 = [
  "torch>=2.7.0",
  "torchvision>=0.22.0"
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", extra = "cu128" }
]
torchvision = [
  { index = "pytorch-cu128", extra = "cu128" }
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
